requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
bcrypt
python-multipart
pydantic-settings
python-dotenv
websockets
